Remove `intel-types` crate
Most of the contents of `intel-types` was moved to the independent `xsave`
crate. For the types that remain, drop `RFlags` and move the others into the
`sgx` crate.
