+++ Makefile.toml
@@ -9,7 +9,6 @@ CARGO_MAKE_CRATE_WORKSPACE_MEMBERS = [
 	"enarx-keep/shims/shim-sgx",
 	"enumerate",
 	"integration-tests",
-	"intel-types",
 	"keep-runtime",
 	"memory",
 	"rcrt1",
