+++ enarx-keep/shims/shim-sgx/Cargo.toml
@@ -11,7 +11,6 @@ crt0stack = { version = "0.1", default-features = false }
 libc = { version = "0.2", features = [] }
 nbytes = "0.1"
 sgx = { path = "../../../sgx", default-features = false }
-intel-types = { path = "../../../intel-types" }
 enumerate = { path = "../../../enumerate" }
 sallyport = { path = "../../../sallyport" }
 bounds = { path = "../../../bounds" }
