--- intel-types/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "intel-types"
-version = "0.1.0"
-authors = ["Lily Sturmann <lsturman@redhat.com>", "Nathaniel McCallum <npmccallum@redhat.com>" ]
-license = "Apache-2.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-enumerate = { path = "../enumerate" }
-bitflags = "1.2.1"
-
-[dev-dependencies]
-memoffset = "0.5.5"
