+++ sgx/Cargo.toml
@@ -16,7 +16,6 @@ asm = ["xsave/asm"]
 [dependencies]
 bounds  = { path = "../bounds" }
 enumerate  = { path = "../enumerate" }
-intel-types = { path = "../intel-types" }
 memory = { path = "../memory" }
 mmap = { path = "../mmap", optional = true }
 openssl = { version = "=0.10.29", features = ["vendored"], optional = true }
