+++ sgx/src/types/sig.rs
@@ -6,7 +6,62 @@
 //! leaf function to verify that the enclave was properly built.
 
 use super::{attr::Attributes, isv, misc::MiscSelect};
-use intel_types::Masked;
+
+use core::fmt::Debug;
+use core::ops::{BitAnd, BitOr, Not};
+
+/// Succinctly describes a masked type, e.g. masked Attributes or masked MiscSelect.
+/// A mask is applied to Attributes and MiscSelect structs in a Signature (SIGSTRUCT)
+/// to specify values of Attributes and MiscSelect to enforce. This struct combines
+/// the struct and its mask for simplicity.
+#[repr(C)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub struct Masked<T: BitAnd<Output = T>> {
+    /// The data being masked, e.g. Attribute flags.
+    pub data: T,
+
+    /// The mask.
+    pub mask: T,
+}
+
+impl<T> Default for Masked<T>
+where
+    T: BitAnd<Output = T>,
+    T: BitOr<Output = T>,
+    T: Not<Output = T>,
+    T: Default,
+    T: Copy,
+{
+    fn default() -> Self {
+        T::default().into()
+    }
+}
+
+impl<T> From<T> for Masked<T>
+where
+    T: BitAnd<Output = T>,
+    T: BitOr<Output = T>,
+    T: Not<Output = T>,
+    T: Copy,
+{
+    fn from(value: T) -> Self {
+        Self {
+            data: value,
+            mask: value | !value,
+        }
+    }
+}
+
+impl<T> PartialEq<T> for Masked<T>
+where
+    T: BitAnd<Output = T>,
+    T: PartialEq,
+    T: Copy,
+{
+    fn eq(&self, other: &T) -> bool {
+        self.mask & self.data == self.mask & *other
+    }
+}
 
 /// The `Author` of an enclave
 ///
