+++ sgx/src/types/ssa.rs
@@ -12,7 +12,6 @@ use core::{
     num::NonZeroU32,
 };
 use enumerate::enumerate;
-use intel_types::*;
 use memory::Register;
 use xsave::XSave;
 
@@ -26,6 +25,71 @@ enumerate! {
         /// Software
         Software = 0b110,
     }
+}
+
+/// Exception Error Codes
+#[repr(u8)]
+#[derive(Copy, Clone, Debug, PartialEq)]
+pub enum Exception {
+    /// Divide-by-zero Error
+    DivideByZero = 0x00,
+
+    /// Debug
+    Debug = 0x01,
+
+    /// Breakpoint
+    Breakpoint = 0x03,
+
+    /// Overflow
+    Overflow = 0x04,
+
+    /// Bound Range Exceeded
+    BoundRange = 0x05,
+
+    /// Invalid Opcode
+    InvalidOpcode = 0x06,
+
+    /// Device Not Available
+    DeviceNotAvailable = 0x07,
+
+    /// Double Fault
+    DoubleFault = 0x08,
+
+    /// Invalid TSS
+    InvalidTss = 0x0A,
+
+    /// Segment Not Present
+    SegmentNotPresent = 0x0B,
+
+    /// Stack-Segment Fault
+    StackSegment = 0x0C,
+
+    /// General Protection Fault
+    GeneralProtection = 0x0D,
+
+    /// Page Fault
+    Page = 0x0E,
+
+    /// x87 Floating-Point Exception
+    FloatingPoint = 0x10,
+
+    /// Alignment Check
+    AlignmentCheck = 0x11,
+
+    /// Machine Check
+    MachineCheck = 0x12,
+
+    /// SIMD Floating-Point Exception
+    SimdFloatingPoint = 0x13,
+
+    /// Virtualization Exception
+    Virtualization = 0x14,
+
+    /// Control Protection Exception
+    ControlProtection = 0x15,
+
+    /// Security Exception
+    Security = 0x1E,
 }
 
 /// Section 38.9.1.1, Table 38-9
@@ -139,7 +203,7 @@ pub struct Gpr {
     pub r15: Register<u64>,
 
     /// Register flags
-    pub rflags: Rflags,
+    pub rflags: Register<u64>,
 
     /// Register rip
     pub rip: Register<u64>,
